created and implemented auth middleware

+++ src/controllers/columnControllers.ts
@@ -4,10 +4,10 @@ import { lastColumnOrder } from '../utils/dataUtils';
 import { Prisma } from '../../generated/prisma';
 
 const createColumn = async (req: Request, res: Response) => {
-  const { name, userId } = req.body;
+  const { name } = req.body;
   //  check if user exists
   const userExists = await prisma.user.findUnique({
-    where: { id: userId },
+    where: { id: req.user.id },
   });
 
   if (!userExists) {
@@ -16,12 +16,12 @@ const createColumn = async (req: Request, res: Response) => {
     });
   }
   //calculate order
-  const newOrder = await lastColumnOrder(userId);
+  const newOrder = await lastColumnOrder(req.user.id);
   //  create column
   const column = await prisma.column.create({
     data: {
       name,
-      userId,
+      userId: req.user.id,
       order: newOrder,
       jobCount: 0,
     },
@@ -33,11 +33,11 @@ const createColumn = async (req: Request, res: Response) => {
   });
 };
 
-const removeColumn = async (req: Request, res: Response) => {
+const removeColumn = async (req: Request<{ id: string }>, res: Response) => {
   // get column from id param
   const column = await prisma.column.findUnique({
     where: {
-      id: req.params.id as string,
+      id: req.params.id,
     },
   });
   //check if column exists
@@ -46,9 +46,16 @@ const removeColumn = async (req: Request, res: Response) => {
       error: 'job item not found',
     });
   }
+  //check if user is authorized
+
+  if (column.userId !== req.user.id) {
+    return res.status(401).json({
+      error: 'Not authorized: invalid user',
+    });
+  }
   //delete column
   const deletedColumn = await prisma.column.delete({
-    where: { id: req.params.id as string },
+    where: { id: req.params.id },
   });
   //send response
   res.status(201).json({
@@ -71,6 +78,12 @@ const updateColumn = async (req: Request<{ id: string }>, res: Response) => {
   if (!column) {
     return res.status(404).json({
       error: 'job item not found',
+    });
+  }
+  //check if user is authorized
+  if (column.userId !== req.user.id) {
+    return res.status(401).json({
+      error: 'Not authorized: invalid user',
     });
   }
   // edit job item
